// frontend-admin/src/api/generated/model/backupArtifactResponseDtoLifecycleState.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export type BackupArtifactResponseDtoLifecycleState = typeof BackupArtifactResponseDtoLifecycleState[keyof typeof BackupArtifactResponseDtoLifecycleState];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BackupArtifactResponseDtoLifecycleState = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;